DRAFT net: service: Add service management support
If service management monitoring is enabled, then the system starts
to generate higher level management events for things happening in
the system.

For example if the network interface gets IPv6 address and its validity
is verified, then CONNECTED event is generated. The CONNECTED event means
that the network interface is usable, and it has IP address. The CONNECTED
event does not necessarily mean that the network connectivity is ok, but there
are other events generated for that. Similarly if all the IP addresses
are removed from the network interface, then the service is disconnected and
a IDLE event is generated as there is no more network connectivity.

This way applications do not need to monitor individual low level network
events happening in the system but can concentrate to monitor events that
are useful for them.

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>