net: http: Call net_send_cb after a timeout in server
We currently terminate the connection from client immediately
after sending the TCP response. This is too fast for the client
which will re-send the final ACK many times. In order to make
the connection shutdown more cleaner, terminate the connection
AFTER we have received ACK to our HTTP response or after a timeout
if the ACK is lost.

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>