drivers: eth: gmac: Enable 802.1Qav support
This commit makes the driver enable HW Qav support, sets the initial
bandwidth shares to 75% - 25%, and implements the set_config api
callback to allow changing the parameters in runtime.

As we only have two hardware queues, we can automatically calculate
the bandwidth share for both based on just one value when setting the
delta Bandwidth.

Signed-off-by: Tomasz Gorochowik <tgorochowik@antmicro.com>