net: tcp2: Refine the sequence number advancement on PSH out
Advance on the sequence number on the amount of data taken
from the send window.

In principal, this can be done in tcp_out(), but it would me more
obscure.

Anyway, let's keep it in the state machine now (until a fresh look
on how the state machine could be further compressed).

Signed-off-by: Oleg Zhurakivskyy <oleg.zhurakivskyy@intel.com>