drivers: modem: gsm: mux: Support GSM 07.10 muxing protocol
Add support to GSM 07.10 muxing protocol which is used to
share the same UART for PPP and AT commands. This allows e.g.
the modem to send SMS and have PPP connection at the same
time.

modem: gsm: Enable PPP interface when connection is ready

By default PPP interface is not taken up automatically but only
after the PPP connection to modem is ready.

drivers: console: uart_mux: Initial UART muxing support

Create support for muxed UARTs which are attached to a real
UART and which use GSM 07.10 muxing protocol to create virtual
channels that can be run on top of the real UART.

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>