net: lib: http_server: Improve POLLERR handling for server sockets
POLLERR shouldn't be ignored, but rather treated as a regular socket
error. Therefore, upon POLLERR on a client socket, close the connection,
and for listening sockets, abort the server operation (ideally, we'd
need to restart the server in such case).

Also, when looping sockets, start with index 1, as we don't need to
monitor eventfd socket in the main loop, it is verified earlier.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>