net: lib: http_server: Fix large HTTP1 payload uploads
Fix HTTP1 request processing when payload exceeds the servers buffer
size. We need to be able to process the payload in fragments.

Make sure we report length == 0 to the application (message end), only
when the whole message has been received, and not at the end of each
processed fragment.

As URL/header fields can also be fragmented, take this into account when
parsing. URL parsing now keeps track of the currently received part.
We only parsed Upgrade header so far, but this information can also be
read from the parser context, so rely on the HTTP1 parser instead.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>