net: lib: http_server: Fix socket count passed to poll()
Keeping track of open sockets doesn't really work well with poll(), as
there may be gaps in the pollfd array (for instance, when two
clients connected, and the first one closed the connection). In effect,
some sockets may not really be monitored resulting in server
malfunction.

This could be addressed by shifting sockets withing pollfd array when
sockets are closed, but since poll() ignores invalid FD anyway, we can
just feed it with the total array size, as the closed socket is
overwritten with a invalid FD in the pollfd array when closed.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>